Module Grpc_stream.Writer

type 'a t = 'a Grpc_eio.Seq.writer

val write : 'a t -> 'a -> unit

Append a value at the current end of the stream.

val close : _ t -> unit

Once all the values have been written to the stream, it must be closed explicitly with this function.